requirements.txt
discord.py
python-dotenv
aiosqlite
playwright
